fix: Add 0 to error metrics on success
This is done so that the error metrics are initialized for that
particular set of labels to 0. By doing this, any backend that receives
the data will receive a time series that starts at 0 instead of
eventually receiving a new time series that starts at 1 once the first
error arrives. In the latter case, a promql query for `rate` e.g. would
fail to detect the transition from 0 -> 1 as the transition is actually
from non-existant to 1.