feat: Enable use of env vars globally in config
The previous implementation hooked in to the same substitution logic as
for step outputs and generated values. This had the consequence that it
would only apply during probes, and can not be used for other parts of
the configuration file such as webhook URLs.

This commit moves the enivronment variable substitution to the
configuration loading step of Prodzilla's initialization in order to
allow environment variables to be used anywhere in the configuration.