chore: Track Cargo.lock
It seems that the guidance used to be that Cargo.lock should be tracked
for binaries (like Prodzilla) but not for libraries. It should have been
 tracked from the start according to this guidance.

 Updated guidance is available at
 https://blog.rust-lang.org/2023/08/29/committing-lockfiles.html , where
 the new recommendation is to simply do what is best for the project but
 defualt to tracking Cargo.lock.

For Prodzilla, tracking Cargo.lock helps make the Docker and binary
builds reproducible so I believe that it is beneficial to this project.

Further reference:
- https://doc.rust-lang.org/nightly/cargo/faq.html#why-have-cargolock-in-version-control
- https://github.com/rust-lang/cargo/issues/8728