fix: Remove openssl vendoring, binary builds
OpenSSL vendoring was causing issues with builds on Windows so I think
for now it's probably best to just remove it and the binary builds that
required it. This means that releases will no longer include
executables, but the Docker images are still published and the source
code can be compiled for the target platform by the user.

It might be worth revisiting binary builds in the future with static
linking, but I don't have the time (or the use-case at $WORK) to do that
at the moment and would like to avoid this being a blocker.